Split validation errors
