+++ libmamba/include/mamba/core/validate.hpp
@@ -78,155 +78,10 @@ namespace mamba::validation
         const std::string& signature
     );
 
-    /**
-     * Base class for artifact/package verification error.
-     */
-    class trust_error : public std::exception
-    {
-    public:
-
-        trust_error(std::string_view message);
-        ~trust_error() override = default;
-        [[nodiscard]] auto what() const noexcept -> const char* override;
-
-    private:
-
-        std::string m_message;
-    };
-
-
-    /**
-     * Error raised when a threshold of signatures is
-     * not met. This can be due to wrong signatures,
-     * wrong or missing public keys.
-     */
-    class threshold_error : public trust_error
-    {
-    public:
-
-        threshold_error();
-        ~threshold_error() override = default;
-    };
-
-    /**
-     * Error raised when wrong metadata are spotted
-     * in a role file.
-     */
-    class role_metadata_error : public trust_error
-    {
-    public:
-
-        role_metadata_error();
-        ~role_metadata_error() override = default;
-    };
-
-
-    /**
-     * Error raised when a wrong file name is
-     * detected for role metadata.
-     */
-    class role_file_error : public trust_error
-    {
-    public:
-
-        role_file_error();
-        ~role_file_error() override = default;
-    };
-
-
-    /**
-     * Error raised when a possible rollback
-     * attack is detected.
-     */
-    class rollback_error : public trust_error
-    {
-    public:
-
-        rollback_error();
-        ~rollback_error() override = default;
-    };
-
-
-    /**
-     * Error raised when a possible freeze
-     * attack is detected.
-     */
-    class freeze_error : public trust_error
-    {
-    public:
-
-        freeze_error();
-        ~freeze_error() override = default;
-    };
-
-
-    /**
-     * Error raised when a spec version is either
-     * wrong/invalid or not supported by the client.
-     */
-    class spec_version_error : public trust_error
-    {
-    public:
-
-        spec_version_error();
-        ~spec_version_error() override = default;
-    };
-
-
-    /**
-     * Error raised when a role metadata file
-     * fetching process fails.
-     */
-    class fetching_error : public trust_error
-    {
-    public:
-
-        fetching_error();
-        ~fetching_error() override = default;
-    };
-
-
-    /**
-     * Error raised when signatures threshold
-     * is not met for a package.
-     */
-    class package_error : public trust_error
-    {
-    public:
-
-        package_error();
-        ~package_error() override = default;
-    };
-
-    /**
-     * Error raised when signatures threshold
-     * is not met for a trust role.
-     */
-    class role_error : public trust_error
-    {
-    public:
-
-        role_error();
-        ~role_error() override = default;
-    };
-
-    /**
-     * Error raised when an invalid package
-     * index is met.
-     */
-    class index_error : public trust_error
-    {
-    public:
-
-        index_error();
-        ~index_error() override = default;
-    };
-
     void check_timestamp_metadata_format(const std::string& ts);
 
     /**
-     * Representation of the public part of a
-     * cryptographic key pair.
+     * Representation of the public part of a cryptographic key pair.
      */
     struct Key
     {
@@ -236,7 +91,7 @@ namespace mamba::validation
 
         static Key from_ed25519(std::string keyval)
         {
-            return { "ed25519", "ed25519", keyval };
+            return { "ed25519", "ed25519", std::move(keyval) };
         }
     };
 
@@ -246,8 +101,8 @@ namespace mamba::validation
 
     /**
      * Representation of a role signature.
-     * Optional 'pgp_trailer' will trigger special
-     * handling during verification to conform to
+     *
+     * Optional 'pgp_trailer' will trigger special handling during verification to conform to
      * OpenPGP RFC4880.
      */
     struct RoleSignature
@@ -265,8 +120,8 @@ namespace mamba::validation
 
     /**
      * Store key IDs and threshold for a role.
-     * Key ID can be a hash of Key, or just
-     * its public key value.
+     *
+     * Key ID can be a hash of Key, or just its public key value.
      */
     struct RoleKeys
     {
@@ -279,8 +134,7 @@ namespace mamba::validation
 
 
     /**
-     * Store key values and threshold for role.
-     * Assumes key scheme/type is `ed25519`.
+     * Store key values and threshold for role. Assumes key scheme/type is `ed25519`.
      */
     struct RolePubKeys
     {
