+++ libmamba/src/core/validate.cpp
@@ -8,7 +8,6 @@
 #include <iostream>
 #include <regex>
 #include <set>
-#include <stdexcept>
 #include <string>
 #include <vector>
 
@@ -22,70 +21,11 @@
 #include "mamba/util/cryptography.hpp"
 #include "mamba/util/string.hpp"
 #include "mamba/util/url.hpp"
+#include "mamba/validation/errors.hpp"
+
 
 namespace mamba::validation
 {
-
-    trust_error::trust_error(std::string_view message)
-        : m_message(util::concat("Content trust error. ", message, ". Aborting."))
-    {
-    }
-
-    auto trust_error::what() const noexcept -> const char*
-    {
-        return this->m_message.c_str();
-    }
-
-    threshold_error::threshold_error()
-        : trust_error("Signatures threshold not met")
-    {
-    }
-
-    role_metadata_error::role_metadata_error()
-        : trust_error("Invalid role metadata")
-    {
-    }
-
-    rollback_error::rollback_error()
-        : trust_error("Possible rollback attack")
-    {
-    }
-
-    freeze_error::freeze_error()
-        : trust_error("Possible freeze attack")
-    {
-    }
-
-    role_file_error::role_file_error()
-        : trust_error("Invalid role file")
-    {
-    }
-
-    spec_version_error::spec_version_error()
-        : trust_error("Unsupported specification version")
-    {
-    }
-
-    fetching_error::fetching_error()
-        : trust_error("Failed to fetch role metadata")
-    {
-    }
-
-    package_error::package_error()
-        : trust_error("Invalid package")
-    {
-    }
-
-    role_error::role_error()
-        : trust_error("Invalid role")
-    {
-    }
-
-    index_error::index_error()
-        : trust_error("Invalid package index metadata")
-    {
-    }
-
     auto sha256sum(const fs::u8path& path) -> std::string_view
     {
         std::ifstream infile = mamba::open_ifstream(path);
