+++ libmamba/tests/src/core/test_validate_v1.cpp
@@ -13,6 +13,7 @@
 #include "mamba/core/fsutil.hpp"
 #include "mamba/core/validate.hpp"
 #include "mamba/util/encoding.hpp"
+#include "mamba/validation/errors.hpp"
 
 #include "mambatests.hpp"
